Primitive Fish and Follow Behavior
added in a very primitive fish with an object pooler that will instantiate 30 fish on load. A very primitive system for the fish to follow the bobber once it is on the water. Fish swim around in a bounding box, can be updated when we have a better mesh too.

NEED TO WORK ON: Hook and hooking mechanics, making fish smarter, camera angle change when fish are getting hooked.

(: