fishing catches fish in the world
fish in the world are now caught.

you cannot tell what fish are attracted to the lure right now, so whatever fish is attracted to the lure just childs itself to the bobber once caught, so i need to fix that.

also added in two scriptable objects, one to create fish, so in the future we can make ones for different fish, and one for the fish data base. Every fish SO created needs to be added to the fish database if its going to be catchable in the area.

Gotta add in different fish pools, and maybe make the fish pools be able to make more than just one type of fish so we dont have like 100 different fish pools all spawning in the different types of fish.