Fish wobble, nip + commenting
Added in a small wobbling shader to make the animation of fish easier once they are made. We can update it to include the UV Map of each fish and just make a new shader for each one (i think, there may be an easier way).

Fish now nip at the line, and stop nipping if the rod is taken in too soon.

commented everything i could for ease of use.

TODO: Fish can nibble at the line and then swim away to steal bait if they're not hungry enough. Want to add bait and trading and a fake trading partner to upgrade things like hook, bait, line, etc.

Have a good week (: